+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and `Removed`.
 ### Fixed
 
 - Multiple error messages are now correctly localized.
+- Corrected error in flavor detection if Ajour was launched before WoW had ever
+  been launched.
 
 ## [0.7.0] - 2021-01-26